Implement an API client for item prices
If we separate the concerns of making API requests and building URLs
into a Class, we can clean up our app.tsx code and follow good
practices. We implement this in such a way that we can
instantiate the API Client and use it for any amount of items, cities or
qualities; this promotes code reuse.